+++ Task_2_1_1/settings.gradle.kts
@@ -10,4 +10,4 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.4.0"
 }
 
-rootProject.name = "parallelPrime"
+rootProject.name = "parallelprime"
