+++ Task_2_1_1/src/jmh/java/ru/nsu/kuklin/parallelprime/OnlyLastCompositeThreadCount.java
@@ -1,13 +1,15 @@
-package ru.nsu.kuklin.parallelPrime;
+package ru.nsu.kuklin.parallelprime;
 
 import java.util.Random;
 import org.openjdk.jmh.annotations.*;
 
 @State(Scope.Benchmark)
-public class OnlyLastComposite {
+public class OnlyLastCompositeThreadCount {
     @Param({ "1000", "10000", "100000", "1000000" })
     public int size;
     public int[] nums;
+    @Param({"1", "2", "4", "8", "16"})
+    public int threadCount;
 
     @Setup(Level.Invocation)
     public void setUp() {
